+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Added limit for asset size and number of concurrent downloads
 - Added kache to the list of known tools
 - Added some tests to the downloader using Go 1.27's new `NewTestServer`
+- Added limit for extracted file size
 
 ### Changed